[i74] fix slider jump after seekTo process

+++ stream-chat-android-ui-common/src/main/kotlin/io/getstream/chat/android/ui/common/state/messages/composer/RecordingState.kt
@@ -9,26 +9,26 @@ public sealed class RecordingState {
     }
 
     public sealed class Recording : RecordingState() {
-        public abstract val duration: Long
+        public abstract val duration: Int
         public abstract val waveform: List<Float>
     }
 
     public data class Hold(
-        override val duration: Long = 0L,
+        override val duration: Int = 0,
         override val waveform: List<Float> = emptyList(),
     ) : Recording() {
         override fun toString(): String = "Recording.Hold(waveform=${waveform.size}, duration=${duration}ms)"
     }
 
     public data class Locked(
-        override val duration: Long = 0L,
+        override val duration: Int = 0,
         override val waveform: List<Float> = emptyList(),
     ) : Recording() {
         override fun toString(): String = "Recording.Locked(waveform=${waveform.size}, duration=${duration}ms)"
     }
 
     public data class Overview(
-        val duration: Long = 0L,
+        val duration: Int = 0,
         val waveform: List<Float> = emptyList(),
         val attachment: Attachment,
         val isPlaying: Boolean = false,
@@ -54,7 +54,7 @@ public sealed class RecordingState {
 }
 
 public fun RecordingState.Recording.copy(
-    duration: Long = this.duration,
+    duration: Int = this.duration,
     waveform: List<Float> = this.waveform,
 ): RecordingState.Recording = when (this) {
     is RecordingState.Hold -> RecordingState.Hold(duration, waveform)
